requirements.txt
boto3==1.24.84
face_recognition==1.3.0
Flask==2.2.2
Flask_Cors==3.0.10
gradio==3.4
moviepy==1.0.3
numpy==1.23.2
pandas==1.4.3
python-dotenv==0.21.0
requests==2.26.0
tensorflow-macos
tensorflow_hub==0.12.0
